undo 1 int

+++ student/yiuchunt_student_turtle.cpp
@@ -143,7 +143,7 @@ void moveTurtle(QPointF &pos_, int8_t &new_orientation, int8_t &current_state,
 // Ground rule -- you are only allowed to call the helper functions "bumped(..)"
 // and "atend(..)", and NO other turtle methods or maze methods (no peeking at
 // the maze!)
-bool studentMoveTurtle(QPointF &pos_, int8_t &new_orientation) {
+bool studentMoveTurtle(QPointF &pos_, int &new_orientation) {
   static const int8_t TIMEOUT =
       4; // bigger number slows down simulation so you can see what's happening
   static const int8_t CYCLE_DECREASE = 1;
